+++ pypgx/cli/plot_bam_read_depth.py
@@ -28,7 +28,8 @@ CovFrame[ReadDepth]."""
         '--path',
         metavar='PATH',
         help=
-"""Create plots in this directory."""
+"""Create plots in this directory (default: current
+directory)."""
     )
     parser.add_argument(
         '--samples',