+++ pypgx/cli/plot_vcf_allele_fraction.py
@@ -28,7 +28,8 @@ VcfFrame[Imported]."""
         '--path',
         metavar='PATH',
         help=
-"""Create plots in this directory."""
+"""Create plots in this directory (default: current
+directory)."""
     )
     parser.add_argument(
         '--samples',